Improving logic around visiting and flattening GroupJoin-SelectMany-DefaultIfEmpty queries.
Previously, when we encountered this type of expression we would produce LEFT JOIN in the SelectExpression but at the same time rewrite the query model, pruning the GJ and DIE clauses, replacing it with regular join clause.
This is problematic, because QM no longer adequately represents the query that is being generated - if the new QM were visited again, we would produce INNER JOIN.

Fix is to not modify the QM, but instead keep track of additional from clauses that were flattened (and not visit them in the VisitAdditionalFromClause), as well as properly remap query/shapers to point to additional from clause, rather than the join.